fix: merge CopyFile func according code review comments with bird
